requirements.txt
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
simple-salesforce>=1.12.0
python-dotenv>=1.0.0
pathlib>=1.0.1
requests>=2.31.0 
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
distro==1.9.0
faiss-cpu==1.9.0
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
isodate==0.7.2
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests-file==2.1.0
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zeep==4.3.1